+++ src/code42cli/cmds/cases.py
@@ -4,7 +4,9 @@ import os
 import click
 from py42.clients.cases import CaseStatus
 from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42CaseAlreadyHasEventError
 from py42.exceptions import Py42NotFoundError
+from py42.exceptions import Py42UpdateClosedCaseError
 
 from code42cli.click_ext.groups import OrderedGroup
 from code42cli.errors import Code42CLIError
@@ -241,6 +243,10 @@ def add(state, case_number, event_id):
     """Associate a file event to a case, by event ID."""
     try:
         state.sdk.cases.file_events.add(case_number, event_id)
+    except Py42UpdateClosedCaseError:
+        raise
+    except Py42CaseAlreadyHasEventError:
+        raise
     except Py42BadRequestError:
         raise Code42CLIError("Invalid case-number or event-id.")